Update app to work with the latest Gandi API
This is in practice a complete rewrite of the application.

Configuration is simplified and the execution can create a
record if it does not exist.

It supports IPv6 if the IP Lookup Service supports it too.

There is no distintion between IPv4 and IPv6 in the configuration,
this should be all abstracted by the application.